Make OsmModule testable
