+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/OsmDatabaseTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueStore;
-import org.opentripplanner.osm.OsmProvider;
+import org.opentripplanner.osm.DefaultOsmProvider;
 import org.opentripplanner.test.support.ResourceLoader;
 
 public class OsmDatabaseTest {
@@ -20,7 +20,7 @@ public class OsmDatabaseTest {
   @Test
   void bicycleRouteRelations() {
     var osmdb = new OsmDatabase(DataImportIssueStore.NOOP);
-    var provider = new OsmProvider(RESOURCE_LOADER.file("ehningen-minimal.osm.pbf"), true);
+    var provider = new DefaultOsmProvider(RESOURCE_LOADER.file("ehningen-minimal.osm.pbf"), true);
     provider.readOsm(osmdb);
     osmdb.postLoad();
 
@@ -40,7 +40,7 @@ public class OsmDatabaseTest {
   void invalidPublicTransportRelation() {
     var osmdb = new OsmDatabase(DataImportIssueStore.NOOP);
     var file = RESOURCE_LOADER.file("brenner-invalid-relation-reference.osm.pbf");
-    var provider = new OsmProvider(file, true);
+    var provider = new DefaultOsmProvider(file, true);
     provider.readOsm(osmdb);
     osmdb.postLoad();
