+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/OsmModuleTest.java
@@ -1,5 +1,6 @@
 package org.opentripplanner.graph_builder.module.osm;
 
+import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -10,18 +11,22 @@ import static org.opentripplanner.street.model.StreetTraversalPermission.ALL;
 import static org.opentripplanner.street.model.StreetTraversalPermission.PEDESTRIAN;
 
 import java.io.File;
+import java.time.ZoneId;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Set;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner._support.time.ZoneIds;
 import org.opentripplanner.astar.model.GraphPath;
 import org.opentripplanner.framework.i18n.LocalizedString;
 import org.opentripplanner.framework.i18n.NonLocalizedString;
+import org.opentripplanner.osm.DefaultOsmProvider;
 import org.opentripplanner.osm.OsmProvider;
 import org.opentripplanner.osm.model.OsmWay;
 import org.opentripplanner.osm.model.OsmWithTags;
+import org.opentripplanner.osm.tagmapping.OsmTagMapper;
 import org.opentripplanner.osm.wayproperty.CreativeNamer;
 import org.opentripplanner.osm.wayproperty.MixinPropertiesBuilder;
 import org.opentripplanner.osm.wayproperty.WayProperties;
@@ -58,7 +63,7 @@ public class OsmModuleTest {
 
     File file = RESOURCE_LOADER.file("map.osm.pbf");
 
-    OsmProvider provider = new OsmProvider(file, true);
+    DefaultOsmProvider provider = new DefaultOsmProvider(file, true);
 
     OsmModule osmModule = OsmModule
       .of(
@@ -123,7 +128,7 @@ public class OsmModuleTest {
     var gg = new Graph(deduplicator);
 
     File file = RESOURCE_LOADER.file("NYC_small.osm.pbf");
-    var provider = new OsmProvider(file, true);
+    var provider = new DefaultOsmProvider(file, true);
     var osmInfoRepository = new DefaultOsmInfoGraphBuildRepository();
     var vehicleParkingRepository = new DefaultVehicleParkingRepository();
     var osmModule = OsmModule
@@ -162,6 +167,60 @@ public class OsmModuleTest {
       }
       edgeEndpoints.add(endpoints);
     }
+  }
+
+
+
+  /**
+   * There is a one-way road which is also marked as a platform in Sky Campus which crashed OSM
+   */
+  @Test
+  void testCrappyOsmPlatform() {
+
+    var provider = new OsmProvider() {
+      @Override
+      public void readOsm(OsmDatabase osmdb) {
+        final OsmWay way = new OsmWay();
+        way.addTag("public_transport", "platform");
+        way.addTag("ref", "123");
+        way.setOsmProvider(this);
+        osmdb.addWay(way);
+      }
+
+      @Override
+      public OsmTagMapper getOsmTagMapper() {
+        return new OsmTagMapper();
+      }
+
+      @Override
+      public void checkInputs() {
+
+      }
+
+      @Override
+      public WayPropertySet getWayPropertySet() {
+        return new WayPropertySet();
+      }
+
+      @Override
+      public ZoneId getZoneId() {
+        return ZoneIds.LONDON;
+      }
+    };
+    var deduplicator = new Deduplicator();
+    var graph = new Graph(deduplicator);
+    var osmInfoRepository = new DefaultOsmInfoGraphBuildRepository();
+    var osmModule = OsmModule
+      .of(
+        provider,
+        graph,
+        osmInfoRepository,
+        new DefaultVehicleParkingRepository()
+      )
+      .withBoardingAreaRefTags(Set.of("naptan:AtcoCode"))
+      .build();
+    osmModule.buildGraph();
+    assertThat(graph.getEdges()).isNotEmpty();
   }
 
   @Test
@@ -322,7 +381,7 @@ public class OsmModuleTest {
     var graph = new Graph(deduplicator);
 
     File file = RESOURCE_LOADER.file("accessno-at-end.pbf");
-    OsmProvider provider = new OsmProvider(file, false);
+    DefaultOsmProvider provider = new DefaultOsmProvider(file, false);
     OsmModule loader = OsmModule
       .of(
         provider,
@@ -348,10 +407,10 @@ public class OsmModuleTest {
   private BuildResult buildParkingLots() {
     var graph = new Graph();
     var service = new DefaultVehicleParkingRepository();
-    var providers = Stream
+    List<OsmProvider> providers = Stream
       .of("B+R.osm.pbf", "P+R.osm.pbf")
       .map(RESOURCE_LOADER::file)
-      .map(f -> new OsmProvider(f, false))
+      .map(f -> (OsmProvider) new DefaultOsmProvider(f, false))
       .toList();
     var module = OsmModule
       .of(providers, graph, new DefaultOsmInfoGraphBuildRepository(), service)
@@ -378,7 +437,7 @@ public class OsmModuleTest {
     var graph = new Graph(deduplicator);
 
     File file = RESOURCE_LOADER.file("usf_area.osm.pbf");
-    var provider = new OsmProvider(file, false);
+    var provider = new DefaultOsmProvider(file, false);
     var osmInfoRepository = new DefaultOsmInfoGraphBuildRepository();
     var vehicleParkingRepository = new DefaultVehicleParkingRepository();
