+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/PlatformLinkerTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.osm.OsmProvider;
+import org.opentripplanner.osm.DefaultOsmProvider;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.service.osminfo.internal.DefaultOsmInfoGraphBuildRepository;
 import org.opentripplanner.service.vehicleparking.internal.DefaultVehicleParkingRepository;
@@ -29,7 +29,7 @@ public class PlatformLinkerTest {
 
     File file = ResourceLoader.of(this).file("skoyen.osm.pbf");
 
-    OsmProvider provider = new OsmProvider(file, false);
+    DefaultOsmProvider provider = new DefaultOsmProvider(file, false);
 
     OsmModule osmModule = OsmModule
       .of(
