+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/TriangleInequalityTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.Test;
 import org.opentripplanner.astar.model.GraphPath;
 import org.opentripplanner.astar.model.ShortestPathTree;
 import org.opentripplanner.model.modes.ExcludeAllTransitFilter;
-import org.opentripplanner.osm.OsmProvider;
+import org.opentripplanner.osm.DefaultOsmProvider;
 import org.opentripplanner.routing.api.request.RequestModes;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.api.request.request.filter.AllowAllTransitFilter;
@@ -51,7 +51,7 @@ public class TriangleInequalityTest {
     graph = new Graph(new Deduplicator());
 
     File file = ResourceLoader.of(TriangleInequalityTest.class).file("NYC_small.osm.pbf");
-    OsmProvider provider = new OsmProvider(file, true);
+    DefaultOsmProvider provider = new DefaultOsmProvider(file, true);
     OsmModule osmModule = OsmModule
       .of(
         provider,
