+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/UnroutableTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.astar.model.GraphPath;
 import org.opentripplanner.astar.model.ShortestPathTree;
-import org.opentripplanner.osm.OsmProvider;
+import org.opentripplanner.osm.DefaultOsmProvider;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.api.request.StreetMode;
 import org.opentripplanner.routing.graph.Graph;
@@ -38,7 +38,7 @@ public class UnroutableTest {
     graph = new Graph(deduplicator);
 
     var osmDataFile = ResourceLoader.of(UnroutableTest.class).file("bridge_construction.osm.pbf");
-    OsmProvider provider = new OsmProvider(osmDataFile, true);
+    DefaultOsmProvider provider = new DefaultOsmProvider(osmDataFile, true);
     OsmModule osmBuilder = OsmModule
       .of(
         provider,
